+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 BUG FIXES
 
 - [core] Fix attempt to double coin deletion
+- [core] Add remainder to total slashed value
 
 ## 0.18.0